requirements.txt
beautifulsoup4
customtkinter
matplotlib
pillow
python-docx
PyYAML
